move static evaluation/analysis into own modules and seperate dir
