+++ lib/Parser.js
@@ -10,6 +10,7 @@ const acorn = require("acorn-dynamic-import").default;
 const traverser = require("acorn/dist/walk");
 const Tapable = require("tapable");
 const BasicEvaluatedExpression = require("./BasicEvaluatedExpression");
+const StaticEvaluator = require("./static-analysis/StaticEvaluator");
 
 function joinRanges(startRange, endRange) {
 	if(!endRange) return startRange;
@@ -34,464 +35,6 @@ const PARSE_OPTIONS_SCRIPT = {
 	sourceType: "script",
 	plugins: {
 		dynamicImport: true
-	}
-};
-
-function joinRanges(startRange, endRange) {
-	if(!endRange) return startRange;
-	if(!startRange) return endRange;
-	return [startRange[0], endRange[1]];
-}
-
-const BinaryExpressionStaticEvaluator = {
-
-	"+": (expression, parser) => {
-		const left = parser.evaluateExpression(expression.left);
-		const right = parser.evaluateExpression(expression.right);
-
-		// probably one of the sides is not static?
-		if(!left || !right) return;
-
-		const res = new BasicEvaluatedExpression();
-		if(left.isString()) {
-			if(right.isString()) {
-				res.setString(left.string + right.string);
-			} else if(right.isNumber()) {
-				res.setString(left.string + right.number);
-			} else if(right.isWrapped() && right.prefix && right.prefix.isString()) {
-				res.setWrapped(
-					new BasicEvaluatedExpression()
-					.setString(left.string + right.prefix.string)
-					.setRange(joinRanges(left.range, right.prefix.range)),
-					right.postfix);
-			} else {
-				res.setWrapped(left, null);
-			}
-		} else if(left.isNumber()) {
-			if(right.isString()) {
-				res.setString(left.number + right.string);
-			} else if(right.isNumber()) {
-				res.setNumber(left.number + right.number);
-			}
-		} else if(left.isWrapped()) {
-			if(left.postfix && left.postfix.isString() && right.isString()) {
-				res.setWrapped(left.prefix,
-					new BasicEvaluatedExpression()
-					.setString(left.postfix.string + right.string)
-					.setRange(joinRanges(left.postfix.range, right.range))
-				);
-			} else if(left.postfix && left.postfix.isString() && right.isNumber()) {
-				res.setWrapped(left.prefix,
-					new BasicEvaluatedExpression()
-					.setString(left.postfix.string + right.number)
-					.setRange(joinRanges(left.postfix.range, right.range))
-				);
-			} else if(right.isString()) {
-				res.setWrapped(left.prefix, right);
-			} else if(right.isNumber()) {
-				res.setWrapped(left.prefix,
-					new BasicEvaluatedExpression()
-					.setString(right.number + "")
-					.setRange(right.range));
-			} else {
-				res.setWrapped(left.prefix, new BasicEvaluatedExpression());
-			}
-		} else {
-			if(right.isString()) {
-				res.setWrapped(null, right);
-			}
-		}
-		res.setRange(expression.range);
-		return res;
-	},
-
-	"-": (expression, parser) => {
-		const left = parser.evaluateExpression(expression.left);
-		const right = parser.evaluateExpression(expression.right);
-		if(!left || !right) return;
-		if(!left.isNumber() || !right.isNumber()) return;
-		const res = new BasicEvaluatedExpression();
-		res.setNumber(left.number - right.number);
-		res.setRange(expression.range);
-		return res;
-	},
-
-	"*": (expression, parser) => {
-		const left = parser.evaluateExpression(expression.left);
-		const right = parser.evaluateExpression(expression.right);
-		if(!left || !right) return;
-		if(!left.isNumber() || !right.isNumber()) return;
-		const res = new BasicEvaluatedExpression();
-		res.setNumber(left.number * right.number);
-		res.setRange(expression.range);
-		return res;
-	},
-
-	"/": (expression, parser) => {
-		const left = parser.evaluateExpression(expression.left);
-		const right = parser.evaluateExpression(expression.right);
-		if(!left || !right) return;
-		if(!left.isNumber() || !right.isNumber()) return;
-		const res = new BasicEvaluatedExpression();
-		res.setNumber(left.number / right.number);
-		res.setRange(expression.range);
-		return res;
-	},
-
-	"===": (expression, parser) => {
-		const left = parser.evaluateExpression(expression.left);
-		const right = parser.evaluateExpression(expression.right);
-		if(!left || !right) return;
-		const res = new BasicEvaluatedExpression();
-		res.setRange(expression.range);
-		if(left.isString() && right.isString()) {
-			return res.setBoolean(left.string === right.string);
-		} else if(left.isNumber() && right.isNumber()) {
-			return res.setBoolean(left.number === right.number);
-		} else if(left.isBoolean() && right.isBoolean()) {
-			return res.setBoolean(left.bool === right.bool);
-		}
-	},
-
-	"!==": (expression, parser) => {
-		const left = parser.evaluateExpression(expression.left);
-		const right = parser.evaluateExpression(expression.right);
-		if(!left || !right) return;
-		const res = new BasicEvaluatedExpression();
-		res.setRange(expression.range);
-		if(left.isString() && right.isString()) {
-			return res.setBoolean(left.string !== right.string);
-		} else if(left.isNumber() && right.isNumber()) {
-			return res.setBoolean(left.number !== right.number);
-		} else if(left.isBoolean() && right.isBoolean()) {
-			return res.setBoolean(left.bool !== right.bool);
-		}
-	},
-};
-
-const StringPrototypeCallExpressionEvaluator = {
-	"replace": (expression, param, parser) => {
-		// we can only optimize `.replace` calls on static strings
-		if(!param.isString()) return;
-
-		// if we have more than 2 arguments, we dont know whats happening - skip
-		if(expression.arguments.length !== 2) return;
-
-		const arg1 = parser.evaluateExpression(expression.arguments[0]);
-		// we may only evaluate the result if the first param is a string or a regexp
-		if(!arg1.isString() && !arg1.isRegExp()) return;
-
-		// get the actual value to search for
-		const search = arg1.regExp || arg1.string;
-
-		const arg2 = parser.evaluateExpression(expression.arguments[1]);
-		// the second arg must be a static string
-		if(!arg2.isString()) return;
-		// get the value to replace the search result with
-		const relaceValue = arg2.string;
-		// perform the replacement
-		const replacementString = param.string.replace(search, relaceValue);
-
-		// create an evaluated replacement
-		return new BasicEvaluatedExpression()
-			.setString(replacementString)
-			.setRange(expression.range);
-	},
-
-	"split": (expression, param, parser) => {
-		// we can only optimize `.split` calls on static strings
-		if(!param.isString()) return;
-
-		// we only know what to do if we have 1 argument
-		if(expression.arguments.length !== 1) return;
-
-		const arg = parser.evaluateExpression(expression.arguments[0]);
-
-		// we may only evaluate the result if the first param is a string or a regexp
-		if(!arg.isString() && !arg.isRegExp()) return;
-
-		// get the actual value with which we split the string
-		const splitter = arg.regExp || arg.string;
-
-		// perform splitting
-		const splittedString = param.string.split(splitter);
-
-		// create an evaluated replacement
-		return new BasicEvaluatedExpression()
-			.setArray(splittedString)
-			.setRange(expression.range);
-	},
-
-	"substring": (expression, param, parser) => {
-		// we can only optimize `.substring` calls on static strings
-		if(!param.isString()) return;
-
-		// we only handle 1 or 2 arguments
-		const argumentsLength = expression.arguments.length;
-		if(argumentsLength < 1 || argumentsLength > 2) {
-			return;
-		}
-
-		const arg1 = parser.evaluateExpression(expression.arguments[0]);
-		// argument 1 must be a number, otherwise we cant handle this
-		if(!arg1.isNumber()) return;
-
-		// we only have one argument?
-		// replace and return evaluated expression
-		if(argumentsLength === 1) {
-			const result = param.string.substring(arg1.number);
-			return new BasicEvaluatedExpression()
-				.setString(result)
-				.setRange(expression.range);
-		}
-
-		const arg2 = parser.evaluateExpression(expression.arguments[1]);
-		// argument 2 must also be a number, otherwise we cant handle this
-		if(!arg2.isNumber()) return;
-
-		// replace and return evaluated expression
-		const result = param.string.substring(arg1.number, arg2.number);
-		return new BasicEvaluatedExpression().setString(result).setRange(expression.range);
-	},
-
-	"substr": (expression, param, parser) => {
-		// we can only optimize `.substr` calls on static strings
-		if(!param.isString()) return;
-
-		// we only handle 1 or 2 arguments
-		const argumentsLength = expression.arguments.length;
-		if(argumentsLength < 1 || argumentsLength > 2) {
-			return;
-		}
-
-		const arg1 = parser.evaluateExpression(expression.arguments[0]);
-		// argument 1 must be a number, otherwise we cant handle this
-		if(!arg1.isNumber()) return;
-
-		// we only have one argument?
-		// replace and return evaluated expression
-		if(argumentsLength === 1) {
-			const result = param.string.substr(arg1.number);
-			return new BasicEvaluatedExpression()
-				.setString(result)
-				.setRange(expression.range);
-		}
-
-		const arg2 = parser.evaluateExpression(expression.arguments[1]);
-		// argument 2 must also be a number, otherwise we cant handle this
-		if(!arg2.isNumber()) return;
-
-		// replace and return evaluated expression
-		const result = param.string.substr(arg1.number, arg2.number);
-		return new BasicEvaluatedExpression().setString(result).setRange(expression.range);
-	},
-};
-
-/**
- * @param {string} kind "cooked" | "raw"
- * @param {any[]} quasis quasis
- * @param {any[]} expressions expressions
- * @param {any} parser parser
- * @return {BasicEvaluatedExpression[]} Simplified template
- */
-function getSimplifiedTemplateResult(kind, quasis, expressions, parser) {
-	var i = 0;
-	var parts = [];
-
-	for(i = 0; i < quasis.length; i++) {
-		parts.push(new BasicEvaluatedExpression().setString(quasis[i].value[kind]).setRange(quasis[i].range));
-
-		if(i > 0) {
-			var prevExpr = parts[parts.length - 2],
-				lastExpr = parts[parts.length - 1];
-			var expr = parser.evaluateExpression(expressions[i - 1]);
-			if(!(expr.isString() || expr.isNumber())) continue;
-
-			prevExpr.setString(prevExpr.string + (expr.isString() ? expr.string : expr.number) + lastExpr.string);
-			prevExpr.setRange([prevExpr.range[0], lastExpr.range[1]]);
-			parts.pop();
-		}
-	}
-	return parts;
-}
-
-const StaticEvaluationVisitor = {
-
-	"Literal": (expression) => {
-		const value = expression.value;
-		switch(true) {
-			case typeof value === "number":
-				return new BasicEvaluatedExpression().setNumber(expression.value).setRange(expression.range);
-			case typeof value === "string":
-				return new BasicEvaluatedExpression().setString(expression.value).setRange(expression.range);
-			case typeof value === "boolean":
-				return new BasicEvaluatedExpression().setBoolean(expression.value).setRange(expression.range);
-			case value === null:
-				return new BasicEvaluatedExpression().setNull().setRange(expression.range);
-			case value instanceof RegExp:
-				return new BasicEvaluatedExpression().setRegExp(expression.value).setRange(expression.range);
-		}
-	},
-
-	/*
-	 * Optimize code by preevualuating static Logical expressions
-	 * e.g.:
-	 *  - Logical and: evaluate false && something() => false
-	 *  - Logical or: evaluate true || something() => true
-	 */
-	"LogicalExpression": (expression, parser) => {
-		const operator = expression.operator;
-		// stop if the operator is not && or ||
-		const isAndOrOr = operator === "&&" || operator === "||";
-		if(!isAndOrOr) {
-			return;
-		}
-
-		const left = parser.evaluateExpression(expression.left);
-		const leftAsBool = left && left.asBool();
-
-		/**
-		 * We can only optimize if the left hand side as a static boolean
-		 */
-		if(typeof leftAsBool !== "boolean") {
-			return;
-		}
-
-		/**
-		 * The left hand side is a boolean. Now check if we can break this expression early.
-		 * -> in case of "&&" a "false" left hand "breaks" the expression early
-		 * -> in case of "||" a "true" left hand "breaks" the expression early
-		 */
-		const canExitEarlyAnd = operator === "&&" && !leftAsBool;
-		const canExitEarlyOr = operator === "||" && leftAsBool;
-		if(canExitEarlyAnd || canExitEarlyOr) {
-			return left.setRange(expression.range);
-		}
-
-		/**
-		 * The left hand side did not allow for early exit.
-		 * Therefore we need to use the right hand side to determine the outcome.
-		 */
-		const right = parser.evaluateExpression(expression.right);
-		return right.setRange(expression.range);
-	},
-
-	"BinaryExpression": (expression, parser) => {
-		/**
-		 * If operator is "==" or "!=" turn them into "===" and "!==" respectively
-		 */
-		const normalizeOperator =
-			expression.operator === "==" ? "===" :
-			expression.operator === "!=" ? "!==" :
-			expression.operator;
-
-		if(!BinaryExpressionStaticEvaluator.hasOwnProperty(normalizeOperator)) {
-			return;
-		}
-
-		return BinaryExpressionStaticEvaluator[normalizeOperator](expression, parser);
-	},
-
-	"ConditionalExpression": (expression, parser) => {
-		const condition = parser.evaluateExpression(expression.test);
-		const conditionValue = condition.asBool();
-
-		// We are able to evaluate the conditiona "test"
-		// use left/right (consequent/alernative) based on the "test" condition
-		if(conditionValue !== undefined) {
-			return parser.evaluateExpression(conditionValue ? expression.consequent : expression.alternate)
-				.setRange(expression.range);
-		}
-
-		// We can not make assumptions about the test
-		// at least try to further optimize the arms of the condition
-		const consequent = parser.evaluateExpression(expression.consequent);
-		const alternate = parser.evaluateExpression(expression.alternate);
-
-		// We need both sides to make assumptions.
-		// Nothing further to optimize here.
-		if(!consequent || !alternate) {
-			return;
-		}
-
-		const res = new BasicEvaluatedExpression();
-		consequent.isConditional() ?
-			res.setOptions(consequent.options) :
-			res.setOptions([consequent]);
-
-		alternate.isConditional() ?
-			res.addOptions(alternate.options) :
-			res.addOptions([alternate]);
-
-		return res.setRange(expression.range);
-	},
-
-	"ArrayExpression": (expression, parser) => {
-		const items = expression.elements
-			.map(element => parser.evaluateExpression(element));
-
-		// We must be able to evaluate every single member
-		// to be able to optimize here.
-		if(!items.every(Boolean)) return;
-
-		return new BasicEvaluatedExpression().setItems(items).setRange(expression.range);
-	},
-
-	/*
-	 * Optimize calls to prototype of certain types
-	 * e.g.
-	 *  - string replace - "asdbar".replace("asd", "foo") => "foobar"
-	 * etc.
-	 *
-	 * check CallExpressionPropertyVisitor for implementation
-	 */
-	"CallExpression": (expression, parser) => {
-		if(expression.callee.type !== "MemberExpression") return;
-		if(expression.callee.property.type !== (expression.callee.computed ? "Literal" : "Identifier")) return;
-		const param = parser.evaluateExpression(expression.callee.object);
-		if(!param) return;
-		const property = expression.callee.property.name || expression.callee.property.value;
-		if(StringPrototypeCallExpressionEvaluator.hasOwnProperty(property)) {
-			return StringPrototypeCallExpressionEvaluator[property](expression, param, parser);
-		}
-		return parser.applyPluginsBailResult("evaluate CallExpression ." + property, expression, param);
-	},
-
-	"TemplateLiteral": (node, parser) => {
-		var parts = getSimplifiedTemplateResult("cooked", node.quasis, node.expressions, parser);
-		if(parts.length === 1) {
-			return parts[0].setRange(node.range);
-		}
-		return new BasicEvaluatedExpression().setTemplateString(parts).setRange(node.range);
-	},
-
-	"TaggedTemplateExpression": (node, parser) => {
-		if(parser.evaluateExpression(node.tag).identifier !== "String.raw") return;
-		var parts = getSimplifiedTemplateResult("raw", node.quasi.quasis, node.quasi.expressions, parser);
-		return new BasicEvaluatedExpression().setTemplateString(parts).setRange(node.range);
-	},
-
-	"UnaryExpression": (expression, parser) => {
-		if(expression.operator === "typeof") {
-			if(expression.argument.type === "FunctionExpression") {
-				return new BasicEvaluatedExpression().setString("function").setRange(expression.range);
-			}
-			let arg = parser.evaluateExpression(expression.argument);
-			if(arg.isString() || arg.isWrapped()) return new BasicEvaluatedExpression().setString("string").setRange(expression.range);
-			else if(arg.isNumber()) return new BasicEvaluatedExpression().setString("number").setRange(expression.range);
-			else if(arg.isBoolean()) return new BasicEvaluatedExpression().setString("boolean").setRange(expression.range);
-			else if(arg.isArray() || arg.isConstArray() || arg.isRegExp()) return new BasicEvaluatedExpression().setString("object").setRange(expression.range);
-		} else if(expression.operator === "!") {
-			let argument = parser.evaluateExpression(expression.argument);
-			if(!argument) return;
-			if(argument.isBoolean()) {
-				return new BasicEvaluatedExpression().setBoolean(!argument.bool).setRange(expression.range);
-			} else if(argument.isString()) {
-				return new BasicEvaluatedExpression().setBoolean(!argument.string).setRange(expression.range);
-			} else if(argument.isNumber()) {
-				return new BasicEvaluatedExpression().setBoolean(!argument.number).setRange(expression.range);
-			}
-		}
 	}
 };
 
@@ -584,39 +127,22 @@ class Parser extends Tapable {
 	}
 
 	initializeEvaluating() {
-		this.plugin("evaluate Literal", function(expr) {
-			return StaticEvaluationVisitor[expr.type](expr);
-		});
-		this.plugin("evaluate LogicalExpression", function(expr) {
-			return StaticEvaluationVisitor[expr.type](expr, this);
-		});
-		this.plugin("evaluate BinaryExpression", function(expr) {
-			return StaticEvaluationVisitor[expr.type](expr, this);
-		});
-		this.plugin("evaluate UnaryExpression", function(expr) {
-			return StaticEvaluationVisitor[expr.type](expr, this);
-		});
+		const boundStaticEvaluator = (expr) => StaticEvaluator.evaluate(expr, this);
+		this.plugin("evaluate Literal", boundStaticEvaluator);
+		this.plugin("evaluate LogicalExpression", boundStaticEvaluator);
+		this.plugin("evaluate BinaryExpression", boundStaticEvaluator);
+		this.plugin("evaluate UnaryExpression", boundStaticEvaluator);
 		this.evaluateTypeOfPlugin.plugin("evaluate typeof undefined", function(expr) {
 			return new BasicEvaluatedExpression().setString("undefined").setRange(expr.range);
 		});
-		this.plugin("evaluate CallExpression", function(expr) {
-			return StaticEvaluationVisitor[expr.type](expr, this);
-		});
+		this.plugin("evaluate CallExpression", boundStaticEvaluator);
 
-		this.plugin("evaluate TemplateLiteral", function(node) {
-			return StaticEvaluationVisitor[node.type](node, this);
-		});
+		this.plugin("evaluate TemplateLiteral", boundStaticEvaluator);
 
-		this.plugin("evaluate TaggedTemplateExpression", function(node) {
-			return StaticEvaluationVisitor[node.type](node, this);
-		});
+		this.plugin("evaluate TaggedTemplateExpression", boundStaticEvaluator);
 
-		this.plugin("evaluate ConditionalExpression", function(expr) {
-			return StaticEvaluationVisitor[expr.type](expr, this);
-		});
-		this.plugin("evaluate ArrayExpression", function(expr) {
-			return StaticEvaluationVisitor[expr.type](expr, this);
-		});
+		this.plugin("evaluate ConditionalExpression", boundStaticEvaluator);
+		this.plugin("evaluate ArrayExpression", boundStaticEvaluator);
 	}
 
 	nameInCurrentScope(name) {
